refactored the Byline component and Story to TS
